TA changed the milliseconds to seconds and changed double to int --> we dont need a floating point number for time values since we just calculate with whole seconds

+++ CombinationsTest/CombinationsTest/Kategorie.cs
@@ -9,11 +9,11 @@ namespace CombinationsTest
     class Kategorie
     {
         private String name;
-        private double usedTime;
-        private double maxTime;
+        private int usedTime;
+        private int maxTime;
         private List<Programm> programs;
 
-        public Kategorie(String n, double ut, double mt, List<Programm> p)
+        public Kategorie(String n, int ut, int mt, List<Programm> p)
         {
             name = n;
             usedTime = ut;
